Extracted boolean decomposition functions from automata_utils.py
