+++ tests/test_boolean_decompose_enfa.py
@@ -7,7 +7,7 @@ from pyformlang.finite_automaton import (
 )
 from scipy.sparse import coo_matrix
 
-from project.automata_utils import (
+from project.boolean_decomposition import (
     boolean_decompose_enfa,
 )
 from project.boolean_decomposition import BooleanDecomposition
